+++ src/main/kotlin/com/github/insanusmokrassar/TelegramBotAPI/bot/Ktor/KtorRequestsExecutor.kt
@@ -60,7 +60,7 @@ class KtorRequestsExecutor(
             if (call == null) {
                 throw IllegalArgumentException("Can't execute request: $request")
             }
-            val content = call.response.content.toByteArray().toString(Charset.defaultCharset())
+            val content = call.response.content.toByteArray().toString(Charsets.UTF_8)
             val responseObject = jsonFormatter.parse(
                 Response.serializer(request.resultSerializer()),
                 content