+++ src/main/kotlin/com/github/insanusmokrassar/TelegramBotAPI/types/buttons/InlineKeyboardMarkup.kt
@@ -1,16 +1,12 @@
 package com.github.insanusmokrassar.TelegramBotAPI.types.buttons
 
 import com.github.insanusmokrassar.TelegramBotAPI.types.buttons.InlineKeyboardButtons.InlineKeyboardButton
+import com.github.insanusmokrassar.TelegramBotAPI.types.inlineKeyboardField
 import kotlinx.serialization.*
 import kotlinx.serialization.internal.ArrayListSerializer
 
 @Serializable
 data class InlineKeyboardMarkup(
-    @SerialName("inline_keyboard")
-    @Serializable(with = KeyboardSerializer::class)
+    @SerialName(inlineKeyboardField)
     val keyboard: Matrix<InlineKeyboardButton>
 ) : KeyboardMarkup
-
-object KeyboardSerializer : KSerializer<Matrix<InlineKeyboardButton>> by ArrayListSerializer(
-    ArrayListSerializer(ContextSerializer(InlineKeyboardButton::class))
-)