+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ## 0.14.0
 
+* Now library have no default engine for both webhooks and requests executor. It is required for clients to set
+some default library
+* All proxy help methods was removed . They are will be replaced in separated project
+* `Ktor` version `1.1.3` -> `1.1.4`
+
 ## 0.13.0 Telegram Polls
 
 * Type `PollOption` and `AnonymousPollOption` added