+++ CHANGELOG.md
@@ -9,6 +9,7 @@ some default library
 * Requests results now always decoding as `UTF-8`
 * `AbstractRequestCallFactory` was added with cache of methods urls to avoid memory leaks
 * Small refactoring of work with response in `KtorRequestsExecutor`
+* Kotlin version `1.3.30` -> `1.3.31`
 
 ## 0.13.0 Telegram Polls